requirements.txt

streamlit==1.37.0
openai==1.30.1  # Groq-compatible
python-dotenv==1.0.1
requests==2.31.0
tensorflow>=2.12.0
pillow>=9.5.0
python-dotenv>=1.0.0

numpy~=1.26.4
transformers==4.41.1
torch>=2.0

